Allow same mac-address vm migration
The problem is when user migrate their vm and want to keep the same mac
address, they specify the same mac address in the target vm. That will
cause kubemacpool manager to reject the target vm because of the mac
address conflict.

The fix is to allow the mac addres confliction when the vm is for
migration purposes. A func is added to check of the conflicting
vms are a migration pair by checking if one of them has runStratgy of
WaitAsReceiver.